+++ app/src/main/java/com/mapbox/maps/testapp/examples/LocationComponentActivity.kt
@@ -100,6 +100,16 @@ class LocationComponentActivity : AppCompatActivity() {
         }
         return true
       }
+      R.id.action_accuracy_enabled -> {
+        binding.mapView.location2.updateSettings2 { showAccuracyRing = true }
+        item.isChecked = true
+        return true
+      }
+      R.id.action_accuracy_disable -> {
+        binding.mapView.location2.updateSettings2 { showAccuracyRing = false }
+        item.isChecked = true
+        return true
+      }
       else -> return super.onOptionsItemSelected(item)
     }
   }