+++ plugin-locationcomponent/src/main/java/com/mapbox/maps/plugin/locationcomponent/LocationLayerRenderer.kt
@@ -18,7 +18,7 @@ internal interface LocationLayerRenderer {
 
   fun show()
 
-  fun styleAccuracy(accuracyAlpha: Float, accuracyColor: Int)
+  fun styleAccuracy(accuracyColor: Int, accuracyBorderColor: Int)
 
   fun setLatLng(latLng: Point)