+++ plugin-locationcomponent/src/main/java/com/mapbox/maps/plugin/locationcomponent/animators/PuckAnimatorManager.kt
@@ -4,46 +4,59 @@ import android.animation.ValueAnimator
 import androidx.annotation.VisibleForTesting
 import androidx.annotation.VisibleForTesting.PRIVATE
 import com.mapbox.geojson.Point
+import com.mapbox.maps.plugin.locationcomponent.*
 import com.mapbox.maps.plugin.locationcomponent.LocationLayerRenderer
-import com.mapbox.maps.plugin.locationcomponent.OnIndicatorBearingChangedListener
-import com.mapbox.maps.plugin.locationcomponent.OnIndicatorPositionChangedListener
 import com.mapbox.maps.plugin.locationcomponent.generated.LocationComponentSettings
+import com.mapbox.maps.plugin.locationcomponent.generated.LocationComponentSettings2
 import com.mapbox.maps.util.MathUtils
 
 internal class PuckAnimatorManager(
   indicatorPositionChangedListener: OnIndicatorPositionChangedListener,
-  indicatorBearingChangedListener: OnIndicatorBearingChangedListener
+  indicatorBearingChangedListener: OnIndicatorBearingChangedListener,
+  indicatorAccuracyRadiusChangedListener: OnIndicatorAccuracyRadiusChangedListener
 ) {
 
   private var bearingAnimator = PuckBearingAnimator(indicatorBearingChangedListener)
   private var positionAnimator = PuckPositionAnimator(indicatorPositionChangedListener)
+  private var accuracyRadiusAnimator =
+    PuckAccuracyRadiusAnimator(indicatorAccuracyRadiusChangedListener)
   private var pulsingAnimator = PuckPulsingAnimator()
 
   @VisibleForTesting(otherwise = PRIVATE)
   constructor(
     indicatorPositionChangedListener: OnIndicatorPositionChangedListener,
     indicatorBearingChangedListener: OnIndicatorBearingChangedListener,
+    indicatorAccuracyRadiusChangedListener: OnIndicatorAccuracyRadiusChangedListener,
     bearingAnimator: PuckBearingAnimator,
     positionAnimator: PuckPositionAnimator,
-    pulsingAnimator: PuckPulsingAnimator
-  ) : this(indicatorPositionChangedListener, indicatorBearingChangedListener) {
+    pulsingAnimator: PuckPulsingAnimator,
+    radiusAnimator: PuckAccuracyRadiusAnimator
+  ) : this(
+    indicatorPositionChangedListener,
+    indicatorBearingChangedListener,
+    indicatorAccuracyRadiusChangedListener
+  ) {
     this.bearingAnimator = bearingAnimator
     this.positionAnimator = positionAnimator
     this.pulsingAnimator = pulsingAnimator
+    this.accuracyRadiusAnimator = radiusAnimator
   }
 
   fun setLocationLayerRenderer(renderer: LocationLayerRenderer) {
     bearingAnimator.setLocationLayerRenderer(renderer)
     positionAnimator.setLocationLayerRenderer(renderer)
     pulsingAnimator.setLocationLayerRenderer(renderer)
+    accuracyRadiusAnimator.setLocationLayerRenderer(renderer)
   }
 
   fun setUpdateListeners(
     onLocationUpdated: ((Point) -> Unit),
-    onBearingUpdated: ((Double) -> Unit)
+    onBearingUpdated: ((Double) -> Unit),
+    onAccuracyRadiusUpdated: ((Double) -> Unit)
   ) {
     positionAnimator.setUpdateListener(onLocationUpdated)
     bearingAnimator.setUpdateListener(onBearingUpdated)
+    accuracyRadiusAnimator.setUpdateListener(onAccuracyRadiusUpdated)
   }
 
   fun onStart() {
@@ -56,13 +69,17 @@ internal class PuckAnimatorManager(
     bearingAnimator.cancelRunning()
     positionAnimator.cancelRunning()
     pulsingAnimator.cancelRunning()
+    accuracyRadiusAnimator.cancelRunning()
   }
 
   fun animateBearing(
     vararg targets: Double,
     options: (ValueAnimator.() -> Unit)?
   ) {
-    bearingAnimator.animate(*MathUtils.prepareOptimalBearingPath(targets).toTypedArray(), options = options)
+    bearingAnimator.animate(
+      *MathUtils.prepareOptimalBearingPath(targets).toTypedArray(),
+      options = options
+    )
   }
 
   fun animatePosition(
@@ -70,6 +87,21 @@ internal class PuckAnimatorManager(
     options: (ValueAnimator.() -> Unit)?
   ) {
     positionAnimator.animate(*targets, options = options)
+  }
+
+  fun animateAccuracyRadius(
+    vararg targets: Double,
+    options: (ValueAnimator.() -> Unit)?
+  ) {
+    accuracyRadiusAnimator.animate(*targets.toTypedArray(), options = options)
+  }
+
+  fun applyAccuracyRadiusSettings(accuracyRadiusSettings: LocationComponentSettings2) {
+    accuracyRadiusAnimator.apply {
+      enabled = accuracyRadiusSettings.showAccuracyRing
+      accuracyCircleColor = accuracyRadiusSettings.accuracyRingColor
+      accuracyCircleBorderColor = accuracyRadiusSettings.accuracyRingBorderColor
+    }
   }
 
   fun applyPulsingAnimationSettings(settings: LocationComponentSettings) {
@@ -91,5 +123,9 @@ internal class PuckAnimatorManager(
 
   fun updatePositionAnimator(block: ValueAnimator.() -> Unit) {
     positionAnimator.updateOptions(block)
+  }
+
+  fun updateAccuracyRadiusAnimator(block: ValueAnimator.() -> Unit) {
+    accuracyRadiusAnimator.updateOptions(block)
   }
 }