+++ plugin-locationcomponent/src/test/java/com/mapbox/maps/plugin/locationcomponent/LocationPuckManagerTest.kt
@@ -12,6 +12,7 @@ import com.mapbox.maps.plugin.delegates.MapCameraManagerDelegate
 import com.mapbox.maps.plugin.delegates.MapDelegateProvider
 import com.mapbox.maps.plugin.locationcomponent.animators.PuckAnimatorManager
 import com.mapbox.maps.plugin.locationcomponent.generated.LocationComponentSettings
+import com.mapbox.maps.plugin.locationcomponent.generated.LocationComponentSettings2
 import com.mapbox.maps.util.captureVararg
 import io.mockk.CapturingSlot
 import io.mockk.every
@@ -32,6 +33,7 @@ import org.robolectric.RobolectricTestRunner
 class LocationPuckManagerTest {
 
   private val settings = mockk<LocationComponentSettings>(relaxed = true)
+  private val accuracyRadiusSettings = mockk<LocationComponentSettings2>(relaxed = true)
   private val delegateProvider = mockk<MapDelegateProvider>(relaxed = true)
   private val mapCameraDelegate = mockk<MapCameraManagerDelegate>(relaxed = true)
   private val style = mockk<StyleInterface>(relaxed = true)
@@ -61,6 +63,7 @@ class LocationPuckManagerTest {
     every { settings.enabled } returns true
     locationPuckManager = LocationPuckManager(
       settings,
+      accuracyRadiusSettings,
       delegateProvider,
       positionManager,
       layerSourceProvider,
@@ -73,7 +76,7 @@ class LocationPuckManagerTest {
   fun testInitialise() {
     locationPuckManager.initialize(style)
     verify { animationManager.setLocationLayerRenderer(locationLayerRenderer) }
-    verify { animationManager.setUpdateListeners(any(), any()) }
+    verify { animationManager.setUpdateListeners(any(), any(), any()) }
     verify { animationManager.applyPulsingAnimationSettings(settings) }
     verify { locationLayerRenderer.addLayers(positionManager) }
     verify { locationLayerRenderer.initializeComponents(style) }
@@ -85,7 +88,7 @@ class LocationPuckManagerTest {
     locationPuckManager.lastLocation = Point.fromLngLat(0.0, 0.0)
     locationPuckManager.initialize(style)
     verify { animationManager.setLocationLayerRenderer(locationLayerRenderer) }
-    verify { animationManager.setUpdateListeners(any(), any()) }
+    verify { animationManager.setUpdateListeners(any(), any(), any()) }
     verify { animationManager.applyPulsingAnimationSettings(settings) }
     verify { locationLayerRenderer.addLayers(positionManager) }
     verify { locationLayerRenderer.initializeComponents(style) }
@@ -267,7 +270,7 @@ class LocationPuckManagerTest {
     locationPuckManager.initialize(style)
 
     verify(exactly = 0) { animationManager.setLocationLayerRenderer(locationLayerRenderer) }
-    verify(exactly = 0) { animationManager.setUpdateListeners(any(), any()) }
+    verify(exactly = 0) { animationManager.setUpdateListeners(any(), any(), any()) }
     verify(exactly = 0) { animationManager.applyPulsingAnimationSettings(settings) }
     verify(exactly = 0) { locationLayerRenderer.addLayers(positionManager) }
     verify(exactly = 0) { locationLayerRenderer.initializeComponents(style) }