+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 Mapbox welcomes participation and contributions from everyone.
 
+## Features ✨ and improvements 🏁
+* Add accuracy radius support for LocationComponent. ([#1016](https://github.com/mapbox/mapbox-maps-android/pull/1016))
+
 # 10.4.0-beta.1
 
 ## Features ✨ and improvements 🏁